Use preferences to identify provided bindings
